deployment/requirements.txt
# Core Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0

# Date/Time Handling
python-dateutil>=2.8.2

# HTTP Requests
requests>=2.31.0
httpx>=0.25.0

# API Documentation
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# OpenAI Integration (if needed)
openai>=1.0.0

# Email Processing (if needed)
email-validator>=2.0.0

# Data Validation
jsonschema>=4.19.0

# File Handling
python-multipart>=0.0.6

# Utilities
python-json-logger>=2.0.7
pytz>=2023.3

# CORS Support
starlette>=0.27.0

# Additional Streamlit Components (optional but useful)
streamlit-calendar>=0.6.0
streamlit-extras>=0.3.0

spotipy>=2.27.0
icalendar>=4.0.9
dateparser>=1.1.8